requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.104.1
h11==0.16.0
httptools==0.6.4
idna==3.10
motor==3.3.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
